wrap App in a ClientOnly guard
