+++ src/app/ClientOnly.tsx
@@ -1,0 +1,14 @@
+import { useEffect, useState } from 'react'
+
+interface ClientOnlyProps {
+  children: React.ReactNode
+  fallback?: React.ReactNode
+}
+
+export function ClientOnly({ children, fallback = null }: ClientOnlyProps) {
+  const [isClient, setIsClient] = useState(false)
+  useEffect(() => {
+    setIsClient(true)
+  }, [])
+  return isClient ? children : fallback || <div className="loading">Loading...</div>
+}
