+++ src/styles.css
@@ -290,3 +290,11 @@ body {
   opacity: 0.7;
   background: var(--code-bg);
 }
+
+.loading {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  width: 100vw;
+}
